Struct cslice::CMutSlice

#[repr(C)]

pub struct CMutSlice<'a, T> { /* fields omitted */ }

A mutable slice, equivalent to &'a mut T.

A CMutSlice can be constructed from a corresponding Rust slice via the AsCMutSlice trait.

A Rust slice can be constructed from a corresponding CMutSlice via as_mut.

Methods

impl<'a, T> CMutSlice<'a, T>

unsafe fn new(base: *mut T, len: usize) -> Self

Create a CSlice from raw data.

Safety

The region of memory from base (inclusive) to base + len * sizeof<T> (exclusive) must be valid for the duration of lifetime 'a.

fn as_ptr(&self) -> *const T

Produces a raw pointer to the slice's buffer.

fn as_mut_ptr(&mut self) -> *mut T

Produces a raw pointer to the slice's buffer.

fn as_slice(&self) -> &'a [T]

A cheap conversion to a Rust slice. This is slightly more general than as_ref.

fn as_mut_slice(&mut self) -> &'a mut [T]

A cheap conversion to a mutable Rust slice. This is slightly more general than as_mut.

fn len(&self) -> usize

Returns the number of elements in the slice.